test(crs): move the xarray-backed NetCDF cases to their own module
`test_crs_less_semantics.py` is a `core` module, so it has to import in the
extras-free pure-wheel run; a module-level `import xarray` aborts collection
there, which the marker-hygiene guard catches. The NetCDF axis-classification
cases need xarray to write their fixtures, so they move to a module that
declares the dependency with `importorskip` and the `xarray` marker.

The mutation check still catches all six mutations across the split.
